Refactor `PaginatedGrid` imports to simplify path structure across components
